ToolLauncher: add a stacked homepage
When starting scopy in the installation folder a new folder called Landing will be created. This folder will contain a index.html file that can be modified by the user at free will.
If this file (index.html) contains text, it will be displayed on the Home page of Scopy in the stacked widget.

Signed-off-by: Daniel Guramulta <Daniel.Guramulta@Analog.com>